feat: Update() in RTGame and ISystem now take float as an argument
The delta_time is now in seconds

feat: erase custom event data when all handlers of the event are done/the event is cancelled
fix: fix RemoveEventHandler

There is now a templated and a non templated version of RemoveEventHandler. The templated one is for events with custom data
and now takes EventHandler as an argument. The non templated override is for SDL events and takes SDL_EventType and EventHandler as arguments.

fix: move template functions in event_system.cpp to the header
refactor: move EventHandler typedef to inside EventSystem
build: set include directory to src

When including headers they will now have to include their paths relative to src. This is done to prevent collisions between stl/library headers and to prevent confusion

chore: rename identifiers throughout the entire source
chore: move class members above methods in header files